minor cleanup and typos (by Basil)
git-svn-id: http://svn.miranda-ng.org/main/trunk@3600 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ plugins/Variables/src/parse_inet.cpp
@@ -32,7 +32,7 @@ static TCHAR *parseUrlEnc(ARGUMENTSINFO *ai) {
 	if (res == NULL) {
 		return NULL;
 	}
-	unsigned int cur = 0;
+	size_t cur = 0;
 	while (cur < strlen(res)) {
 		if (( (*(res+cur) >= '0') && (*(res+cur) <= '9')) || ( (*(res+cur) >= 'a') && (*(res+cur) <= 'z')) || ( (*(res+cur) >= 'A') && (*(res+cur) <= 'Z')) ) {
 			cur++;
